backend/python/requirements.txt
acres==0.5.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comtypes==1.4.11
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.4
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
etelemetry==0.3.1
fastapi==0.115.14
ffmpeg==1.4
filelock==3.18.0
fitz==0.0.1.dev2
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
looseversion==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pdfminer==20191125
pdfminer.six==20250506
pillow==11.2.1
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.1
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.2
pywin32==310
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2